Remove atomics and reduce memory writes
This patch does 2 main optimizations:
* removal of atomic updates
* only do writes when there's an actual data change

GPU atomics are an expensive operation and a limited GPU resource.
The sh_simul shader was using them to flip the D-bit. Since there is already
a memory barried between each simul stage, and since the D bit of a LUT
that is being simulated will only be used at the earliest in the next
simul stage, there is no need for there to be an atomic update. So
atomicOr and atomicAnd has been replaced by first a read and then a write.

In addition, the old code always writes to the D bit even if the D
bit already has the correct value. Memory BW is almost certainly a bigger
bottleneck than shader instruction throughput. So both in sh_simul and
sh_posedge, writes to outputs are now only done if there's a change in value.

Combined, these changes result in a more-than-doubling of the GPU simulation
performance.